Replace content_[un]like regexps with grep_inputs
Tests were failing due to regexp not matching the order.
For example, we expect to have name and then type,
but we get type and then name, which shouldn't be a big
problem, but it makes tests fail. Using grep_inputs and
seeing if there is a field with given properties fixes
the problem, and tests pass just fine.